// src/features/database/services/database-sync-types.ts
export interface MongoProductDoc {
  _id?: ObjectId;
  id?: string;
  sku?: string | null;
  baseProductId?: string | null;
  defaultPriceGroupId?: string | null;
  ean?: string | null;
  gtin?: string | null;
  asin?: string | null;
  name_en?: string | null;
  name_pl?: string | null;
  name_de?: string | null;
  description_en?: string | null;
  description_pl?: string | null;
  description_de?: string | null;
  supplierName?: string | null;
  supplierLink?: string | null;
  priceComment?: string | null;
  stock?: number | null;
  price?: number | null;
  sizeLength?: number | null;
  sizeWidth?: number | null;
  weight?: number | null;
  length?: number | null;
  parameters?: Array<{ parameterId: string; value_en?: string; value_pl?: string; value_de?: string }>;
  imageLinks?: string[];
  imageBase64s?: string[];
  createdAt?: Date;
  updatedAt?: Date;
  images?: Array<{ imageFileId: string; assignedAt?: Date }>;
  catalogs?: Array<{ catalogId: string; assignedAt?: Date }>;
  categoryId?: string | null;
  categories?: Array<{ categoryId: string; assignedAt?: Date }>;
  tags?: Array<{ tagId: string; assignedAt?: Date }>;
  producers?: Array<{ producerId: string; assignedAt?: Date }>;
}